requirements.lock.txt
# This file was autogenerated by uv via the following command:
#    uv pip compile requirements.txt --output-file=requirements.lock.txt
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via
    #   opentelemetry-exporter-otlp-proto-grpc
    #   posthog
bcrypt==4.3.0
    # via chromadb
build==1.2.2.post1
    # via chromadb
cachetools==5.5.2
    # via google-auth
certifi==2025.7.14
    # via
    #   httpcore
    #   httpx
    #   kubernetes
    #   requests
cfgv==3.4.0
    # via pre-commit
charset-normalizer==3.4.2
    # via requests
chromadb==1.0.15
    # via -r requirements.txt
click==8.2.1
    # via
    #   typer
    #   uvicorn
coloredlogs==15.0.1
    # via onnxruntime
deprecated==1.2.18
    # via opentelemetry-api
distlib==0.4.0
    # via virtualenv
distro==1.9.0
    # via posthog
durationpy==0.10
    # via kubernetes
fastapi==0.116.1
    # via -r requirements.txt
filelock==3.18.0
    # via
    #   huggingface-hub
    #   virtualenv
flatbuffers==25.2.10
    # via onnxruntime
fsspec==2025.7.0
    # via huggingface-hub
gitdb==4.0.12
    # via gitpython
gitpython==3.1.44
    # via -r requirements.txt
google-auth==2.40.3
    # via kubernetes
googleapis-common-protos==1.70.0
    # via opentelemetry-exporter-otlp-proto-grpc
grpcio==1.73.1
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
hf-xet==1.1.5
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   chromadb
    #   langsmith
    #   ollama
huggingface-hub==0.33.4
    # via tokenizers
humanfriendly==10.0
    # via coloredlogs
identify==2.6.12
    # via pre-commit
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
importlib-metadata==4.13.0
    # via -r requirements.txt
importlib-resources==6.5.2
    # via chromadb
jsonpatch==1.33
    # via langchain-core
jsonpointer==3.0.0
    # via jsonpatch
jsonschema==4.25.0
    # via chromadb
jsonschema-specifications==2025.4.1
    # via jsonschema
kubernetes==33.1.0
    # via chromadb
langchain==0.3.26
    # via -r requirements.txt
langchain-core==0.3.70
    # via
    #   langchain
    #   langchain-text-splitters
langchain-text-splitters==0.3.8
    # via langchain
langsmith==0.4.8
    # via
    #   langchain
    #   langchain-core
loguru==0.7.3
    # via -r requirements.txt
markdown-it-py==3.0.0
    # via rich
mdurl==0.1.2
    # via markdown-it-py
mmh3==5.1.0
    # via chromadb
mpmath==1.3.0
    # via sympy
nodeenv==1.9.1
    # via pre-commit
numpy==2.3.1
    # via
    #   chromadb
    #   onnxruntime
oauthlib==3.3.1
    # via
    #   kubernetes
    #   requests-oauthlib
ollama==0.5.1
    # via -r requirements.txt
onnxruntime==1.22.1
    # via chromadb
opentelemetry-api==1.16.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
    #   opentelemetry-sdk
opentelemetry-exporter-otlp-proto-grpc==1.16.0
    # via chromadb
opentelemetry-proto==1.16.0
    # via opentelemetry-exporter-otlp-proto-grpc
opentelemetry-sdk==1.16.0
    # via
    #   chromadb
    #   opentelemetry-exporter-otlp-proto-grpc
opentelemetry-semantic-conventions==0.37b0
    # via opentelemetry-sdk
orjson==3.11.0
    # via
    #   chromadb
    #   langsmith
overrides==7.7.0
    # via chromadb
packaging==25.0
    # via
    #   build
    #   huggingface-hub
    #   langchain-core
    #   langsmith
    #   onnxruntime
platformdirs==4.3.8
    # via virtualenv
posthog==5.4.0
    # via chromadb
pre-commit==4.2.0
    # via -r requirements.txt
protobuf==4.25.8
    # via
    #   googleapis-common-protos
    #   onnxruntime
    #   opentelemetry-proto
pyasn1==0.6.1
    # via
    #   pyasn1-modules
    #   rsa
pyasn1-modules==0.4.2
    # via google-auth
pybase64==1.4.1
    # via chromadb
pydantic==2.11.7
    # via
    #   chromadb
    #   fastapi
    #   langchain
    #   langchain-core
    #   langsmith
    #   ollama
pydantic-core==2.33.2
    # via pydantic
pygments==2.19.2
    # via rich
pypika==0.48.9
    # via chromadb
pyproject-hooks==1.2.0
    # via build
python-dateutil==2.9.0.post0
    # via
    #   kubernetes
    #   posthog
python-dotenv==1.1.1
    # via uvicorn
pyyaml==6.0.2
    # via
    #   chromadb
    #   huggingface-hub
    #   kubernetes
    #   langchain
    #   langchain-core
    #   pre-commit
    #   uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
requests==2.32.4
    # via
    #   -r requirements.txt
    #   huggingface-hub
    #   kubernetes
    #   langchain
    #   langsmith
    #   posthog
    #   requests-oauthlib
    #   requests-toolbelt
requests-oauthlib==2.0.0
    # via kubernetes
requests-toolbelt==1.0.0
    # via langsmith
rich==14.0.0
    # via
    #   chromadb
    #   typer
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
rsa==4.9.1
    # via google-auth
setuptools==80.9.0
    # via
    #   opentelemetry-api
    #   opentelemetry-sdk
shellingham==1.5.4
    # via typer
six==1.17.0
    # via
    #   kubernetes
    #   posthog
    #   python-dateutil
smmap==5.0.2
    # via gitdb
sniffio==1.3.1
    # via anyio
sqlalchemy==2.0.41
    # via langchain
starlette==0.47.2
    # via fastapi
sympy==1.14.0
    # via onnxruntime
tenacity==9.1.2
    # via
    #   chromadb
    #   langchain-core
tokenizers==0.21.2
    # via chromadb
tqdm==4.67.1
    # via
    #   chromadb
    #   huggingface-hub
tree-sitter==0.25.0
    # via -r requirements.txt
typer==0.16.0
    # via chromadb
typing-extensions==4.14.1
    # via
    #   anyio
    #   chromadb
    #   fastapi
    #   huggingface-hub
    #   langchain-core
    #   opentelemetry-sdk
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   starlette
    #   typer
    #   typing-inspection
typing-inspection==0.4.1
    # via pydantic
urllib3==2.5.0
    # via
    #   kubernetes
    #   requests
uv==0.8.0
    # via -r requirements.txt
uvicorn==0.35.0
    # via
    #   -r requirements.txt
    #   chromadb
uvloop==0.21.0
    # via uvicorn
virtualenv==20.32.0
    # via pre-commit
watchfiles==1.1.0
    # via uvicorn
websocket-client==1.8.0
    # via kubernetes
websockets==15.0.1
    # via uvicorn
wrapt==1.17.2
    # via deprecated
zipp==3.23.0
    # via importlib-metadata
zstandard==0.23.0
    # via langsmith
